Splitted code base into multiple ones. Refactored code server
