--- Cargo.toml
@@ -1,33 +1,0 @@
-[package]
-name = "tcproxy"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[[bin]]
-name = "server"
-path = "src/bin/server.rs"
-
-[[bin]]
-name = "client"
-path = "src/bin/client.rs"
-
-[profile.release]
-debug = true
-
-[dependencies]
-console-subscriber = "0.1.5"
-tokio = { version = "1.20.1", features = ["full", "tracing"] }
-tokio-util = { version = "0.7.3", features = ["codec"]}
-bytes = "1.2.0"
-tracing = "0.1.35"
-tracing-subscriber = "0.3.15"
-uuid = { version = "1.1.2", features = ["v4"]}
-clap = { version = "3.2.15", features = ["derive", "color", "default"] }
-crc32fast = "1.3.2"
-chrono = "0.4.19"
-tokio-stream = "0.1.9"
-futures-util = "0.3.21"
-rand = "0.8.5"
-async-stream = "0.3.0"
