--- src/codec/mod.rs
@@ -1,2 +1,0 @@
-mod tcp_stream;
-pub use tcp_stream::*;
