--- src/proxy/mod.rs
@@ -1,9 +1,0 @@
-mod proxy_server;
-mod proxy_client;
-mod proxy_client_reader;
-mod proxy_client_writer;
-
-pub use proxy_server::*;
-pub use proxy_client::*;
-pub use proxy_client_reader::ProxyClientStreamReader;
-pub use proxy_client_writer::ProxyClientStreamWriter;
