--- src/tcp/mod.rs
@@ -1,5 +1,0 @@
-mod tcp_connection;
-mod listener;
-
-pub use tcp_connection::*;
-pub use listener::ListenerUtils;
