+++ tcproxy-cli/src/lib.rs
@@ -1,0 +1,7 @@
+mod local_connection;
+mod commands;
+mod client_state;
+
+pub use commands::*;
+pub use local_connection::*;
+pub use client_state::*;
