+++ tcproxy-core/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "tcproxy-core"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[lib]
+name = "tcproxy_core"
+path = "src/lib.rs"
+
+[dependencies]
+bytes = "1.2.0"
+tracing = "0.1.35"
+async-trait = "0.1.57"
+tokio = { version = "1.20.1", features = ["full", "tracing"] }
+uuid = { version = "1.1.2", features = ["v4"]}
