+++ tcproxy-core/src/command.rs
@@ -1,0 +1,8 @@
+use async_trait::async_trait;
+
+use crate::Result;
+
+#[async_trait]
+pub trait Command: Sync + Send {
+    async fn handle(&self) -> Result<()>;
+}
