+++ tcproxy-core/src/frame_error.rs
@@ -1,0 +1,43 @@
+use std::{string::FromUtf8Error, num::TryFromIntError};
+
+#[derive(Debug)]
+pub enum FrameError {
+    Incomplete,
+    Other(crate::Error),
+}
+
+impl From<String> for FrameError {
+  fn from(src: String) -> FrameError {
+      FrameError::Other(src.into())
+  }
+}
+
+impl From<&str> for FrameError {
+  fn from(src: &str) -> FrameError {
+      FrameError::Other(src.into())
+  }
+}
+
+impl std::error::Error for FrameError {}
+
+impl From<FromUtf8Error> for FrameError {
+  fn from(_src: FromUtf8Error) -> FrameError {
+      "protocol error; invalid frame format".into()
+  }
+}
+
+impl From<TryFromIntError> for FrameError {
+  fn from(_src: TryFromIntError) -> FrameError {
+      "protocol error; invalid frame format".into()
+  }
+}
+
+
+impl std::fmt::Display for FrameError {
+  fn fmt(&self, fmt: &mut std::fmt::Formatter) -> std::fmt::Result {
+      match self {
+          FrameError::Incomplete => "stream ended early".fmt(fmt),
+          FrameError::Other(err) => err.fmt(fmt),
+      }
+  }
+}
